a2a-black-swan-agent/src/__main__.py
import logging
import os

import click
import uvicorn

from a2a.server.apps import A2AStarletteApplication
from a2a.server.request_handlers import DefaultRequestHandler
from a2a.server.tasks import InMemoryTaskStore
from a2a.types import (
    AgentCapabilities,
    AgentCard,
    AgentSkill,
)
from dotenv import load_dotenv
from openai_agent import create_agent  # type: ignore[import-not-found]
from openai_agent_executor import (
    OpenAIAgentExecutor,  # type: ignore[import-untyped]
)
from starlette.applications import Starlette
from starlette.middleware import Middleware
from starlette.middleware.cors import CORSMiddleware


load_dotenv()

logging.basicConfig()


@click.command()
@click.option('--host', 'host', default='localhost')
@click.option('--port', 'port', default=5000)
def main(host: str, port: int):
    # Verify an API key is set.
    if not os.getenv('OPENAI_API_KEY'):
        raise ValueError('OPENAI_API_KEY environment variable not set')

    skill = AgentSkill(
        id='stress_test_strategy',
        name='Strategy Robustness Analyzer',
        description='Runs Monte Carlo simulations, noise injection, and synthetic Black Swan events against a provided trading strategy. Returns structured JSON diagnostics.',
        tags=['quantitative-finance', 'stress-testing', 'backtesting', 'monte-carlo'],
        examples=['Stress test an SMA crossover strategy on AAPL for the last 2 years', 'Run a robustness suite on RSI strategy for TSLA with 14-period lookback'],
    )

    # AgentCard for OpenAI-based agent
    agent_card = AgentCard(
        name='black-swan-agent',
        description='Adversarial stress-testing agent for algorithmic trading strategies.',
        url=f'http://{host}:{port}/',
        version='1.0.0',
        default_input_modes=['text'],
        default_output_modes=['text'],
        capabilities=AgentCapabilities(streaming=True),
        skills=[skill],
    )

    # Create OpenAI agent
    agent_data = create_agent()

    agent_executor = OpenAIAgentExecutor(
        card=agent_card,
        tools=agent_data['tools'],
        api_key=os.getenv('OPENAI_API_KEY'),
        system_prompt=agent_data['system_prompt'],
    )

    request_handler = DefaultRequestHandler(
        agent_executor=agent_executor, task_store=InMemoryTaskStore()
    )

    a2a_app = A2AStarletteApplication(
        agent_card=agent_card, http_handler=request_handler
    )
    routes = a2a_app.routes()

    middleware = [
        Middleware(CORSMiddleware, allow_origins=['*'], allow_methods=['*'], allow_headers=['*'])
    ]

    app = Starlette(routes=routes, middleware=middleware)

    uvicorn.run(app, host=host, port=port)


if __name__ == '__main__':
    main()