a2a-black-swan-agent/pyproject.toml
[project]
name = "black-swan-agent"
version = "0.1.0"
description = "Adversarial stress-testing agent for algorithmic trading strategies."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "openai>=1.57.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    # Quant / Math dependencies
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "yfinance>=0.2.30",
    "pandas-ta>=0.3.14b0",
    "optuna>=3.4.0",
    "numba>=0.59.0; python_version < '3.13'",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"